refactor: dto 디렉토리 변경 적용
